week-06/test_project/requirements.txt
Django==4.1.7
Pillow==9.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
